move home and index view to package entry point because they\'re both equally important.
