+++ source/src/main/java/ch/bfh/bti7081/s2020/blue/view/IndexView.java
@@ -1,7 +1,6 @@
-package ch.bfh.bti7081.s2020.blue.view.landing;
+package ch.bfh.bti7081.s2020.blue.view;
 
 import ch.bfh.bti7081.s2020.blue.view.authentication.LoginViewImpl;
-import ch.bfh.bti7081.s2020.blue.view.landing.LandingView.LandingPageViewListener;
 import com.vaadin.flow.component.Text;
 import com.vaadin.flow.component.html.Anchor;
 import com.vaadin.flow.component.html.H1;
@@ -10,11 +9,16 @@ import com.vaadin.flow.component.html.ListItem;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.router.Route;
+import com.vaadin.flow.server.PWA;
 
-@Route("")
-public class LandingViewImpl extends VerticalLayout implements LandingPageViewListener {
+@Route
+@PWA(name = "Social Anxiety Application",
+    shortName = "Social Anxiety App",
+    description = "This is an application to help people with social anxiety.",
+    enableInstallPrompt = false)
+public class IndexView extends VerticalLayout {
 
-  public LandingViewImpl() {
+  public IndexView() {
     H1 title = new H1("Willkommen zur soziale Phobie App");
 
     HorizontalLayout container = new HorizontalLayout();
@@ -55,5 +59,4 @@ public class LandingViewImpl extends VerticalLayout implements LandingPageViewLi
     setSizeFull();
     add(title, container);
   }
-
 }
