--- source/src/main/java/ch/bfh/bti7081/s2020/blue/view/landing/LandingView.java
@@ -1,11 +1,0 @@
-package ch.bfh.bti7081.s2020.blue.view.landing;
-
-public interface LandingView {
-
-  interface LandingPageViewListener {
-
-  }
-
-  void addListener(LandingPageViewListener listener);
-
-}
